requirements.txt
Django==2.2.3
django-environ==0.4.5
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
pytz==2019.1
sqlparse==0.3.0
